workflow_orchestration/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nyc-taxi-trip-data"
dynamic = ["version"]
description = ''
readme = "readme.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Eric Ti Yu Chiang", email = "eric.ti.yu.chiang@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "google-cloud-bigquery>=3.11.4",
  "httpx>=0.25.0",
  "pandas[parquet,gcp]>=2.1.0",
  "prefect>=2.13.4",
  "prefect-gcp>=0.5.0",
]

[project.urls]
Documentation = "https://github.com/unknown/nyc-taxi-trip-data#readme"
Issues = "https://github.com/unknown/nyc-taxi-trip-data/issues"
Source = "https://github.com/unknown/nyc-taxi-trip-data"

[tool.hatch.version]
path = "__init__.py"

[tool.black]
target-version = ["py37"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py37"
line-length = 120
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.isort]
known-first-party = ["nyc_taxi_trip_data"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["nyc_taxi_trip_data", "tests"]
branch = true
parallel = true

[tool.coverage.paths]
nyc_taxi_trip_data = ["src/nyc_taxi_trip_data", "*/nyc-taxi-trip-data/src/nyc_taxi_trip_data"]
tests = ["tests", "*/nyc-taxi-trip-data/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
